Add NewModule to play with hpc
When we don't cover this new module w/tests at all, the coverage report
fails to mention it entirely! Even by highlighting the entire module yellow.

Once we import + use it in a test, the coverage report includes it.

Seems like it should. Maybe a patch or change in config is in order.